+++ src/main/java/com/software/triviabot/repo/IQuestionRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IQuestionRepo extends JpaRepository<Question, Long> {
+public interface IQuestionRepo extends JpaRepository<Question, Integer> {
 }
